Fourth Commit -ManualAddExpenseScreen & ExpenseListScreen are functioning -Bottom Navigation Bar

+++ app/src/main/java/com/example/aiexpenzo/data/model/User.kt
@@ -5,6 +5,6 @@ data class User(
     val name: String,
     val email: String,
     val password: String,
-    val monthlyIncome: Double = 0.0,
-    val monthlyBudget: Double = 0.0
+    val monthlyIncome: Float = 0f,
+    val monthlyBudget: Float = 0f
 )
